chore: export factory from colors
